# server/models/common.go
package models

import (
	"server/models/entity"

	"gorm.io/gorm"
)

type DbInstance struct {
	DB *gorm.DB
}

type DotEnv struct {
	DB_HOST        string
	DB_USER        string
	DB_PASSWORD    string
	DB_NAME        string
	APP_JWT_SECRET string
}

type ApiErrorWrapper struct {
	Message string `json:"message"`
	Ok      bool   `json:"ok"`
}

var Entities = []any{
	&entity.Organization{},
	&entity.Image{},

	&entity.Venue{},
	&entity.Event{},

	&entity.Booking{},
}